feat: join community fund modal
